backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
fastapi==0.109.0
greenlet==1.1.0
h11==0.12.0
httpcore==0.15
idna==3.2
importlib-metadata==4.6.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pydantic==1.8.2
pyhumps==3.0.2
python-dateutil==2.8.2
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.26.0
six==1.16.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing-extensions
urllib3==1.26.6
uvicorn==0.14.0
zipp==3.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-login==1.9.2
minio==7.2.3